planets
Okay so OpenPanel is for clicking on the planets to bring up the planet's panel which I don't think works but I also don't really know how to test it out.
PlanetTrigger is closer to what we wanted cause it should bring up the planet's panel if the rocket "collides" with the planet but the rocket doesn't fly so I also couldnt test this.
They all use SetActive cause I based it on like the pause menu tutorials I saw which was all I could find on switching panels so as it is now the planet panels would have to move ig
idk sorry I tried